modules: create a separate module for the instance
